+++ eve_industry/code/source/item_attribute.h
@@ -34,7 +34,7 @@ class item_attribute_t
         // really need to be.
         
         /// @brief Integral uniqueifier used by the EvE API
-        unsigned id_;
+        unsigned long id_;
         
         /// @brief The name of this item as it appears in the EvE universe, such
         /// as "Tritanium"
@@ -69,7 +69,7 @@ class item_attribute_t
         // Try to use initializer lists when possible.
         
         /// @brief Initialization constructor
-        inline item_attribute_t(unsigned id, std::string_view name)
+        inline item_attribute_t(unsigned long id, std::string_view name)
           : id_(id),
             name_(name)
         {
@@ -88,7 +88,7 @@ class item_attribute_t
         }
         
         /// @brief Return true if the item ID of this object matches a given ID.
-        inline bool operator==(unsigned id) const
+        inline bool operator==(unsigned long id) const
         {
             return this->id_ == id;
         }
@@ -99,7 +99,7 @@ class item_attribute_t
             return !(*this == source);
         }
         
-        inline bool operator!=(unsigned id) const
+        inline bool operator!=(unsigned long id) const
         {
             return !(*this == id);
         }
@@ -109,13 +109,13 @@ class item_attribute_t
         // inline void member(const type& t);
         
         /// @brief Read access to @ref id_.
-        inline unsigned id() const
+        inline unsigned long id() const
         {
             return this->id_;
         }
         
         /// @brief Write access to @ref id_.
-        inline void id(unsigned id)
+        inline void id(unsigned long id)
         {
             this->id_ = id;
         }
@@ -141,11 +141,11 @@ class item_attribute_t
         /// @brief Connect to the EvE API and pull all of the data associated
         /// with a given item id.
         
-        void fetch(unsigned id);
+        void fetch(unsigned long id);
         
-        /// @brief Same as @ref fetch(unsigned id) but re-uses a reader object
+        /// @brief Same as @ref fetch(unsigned long id) but re-uses a reader object
         /// for efficiency.
-        void fetch(unsigned id, Json::CharReader* reader);
+        void fetch(unsigned long id, Json::CharReader* reader);
         
         /// @brief Open a file conforming to schema.json and use it to initialize
         /// this object, clearing previous content.
